Split fs into mods
